FFS-464 wip configuration page
