+++ src/features/configuration/components/IncomingDataComponent.tsx
@@ -1,16 +1,17 @@
-import { Box, Typography } from "@mui/material";
+import {Box, MenuItem, TextField, Typography} from "@mui/material";
 import * as React from "react";
-import { useContext, useEffect, useState } from "react";
+import {useContext, useEffect, useState} from "react";
 import HelpPopover from "./common/popover/HelpPopover";
-import { useTranslation } from "react-i18next";
-import { SourceApplicationContext } from "../../../context/SourceApplicationContext";
+import {useTranslation} from "react-i18next";
+import {SourceApplicationContext} from "../../../context/SourceApplicationContext";
 import {
 	IInstanceMetadataContent,
 	IInstanceObjectCollectionMetadata,
+	IIntegrationMetadata,
 	ValueType,
 } from "../types/Metadata/IntegrationMetadata";
-import { ClassNameMap } from "@mui/styles";
-import { metadataPanelSX } from "../../../util/styles/SystemStyles";
+import {ClassNameMap} from "@mui/styles";
+import {metadataPanelSX} from "../../../util/styles/SystemStyles";
 import {
 	extractCollectionFieldReferenceIndexAndKey,
 	extractFieldReferenceKey,
@@ -18,189 +19,247 @@ import {
 	isFieldReference,
 } from "../util/FieldReferenceUtils";
 import MetadataContentComponent from "./metadata/MetadataContentComponent";
-import { toInstanceFieldReference } from "../../../util/JsonUtil";
+import {toInstanceFieldReference} from "../../../util/JsonUtil";
 import ObjectCollectionMetadataContentComponent from "./metadata/ObjectCollectionMetadataContentComponent";
 import ValueConvertingRepository from "../../../api/ValueConvertingRepository";
-import { Tag } from "./common/dnd/Tag";
-import { IValueConverting } from "../../valueConverting/types/ValueConverting";
+import {Tag} from "./common/dnd/Tag";
+import {IValueConverting} from "../../valueConverting/types/ValueConverting";
+import {IntegrationContext} from "../../../context/IntegrationContext";
+import {useFormContext} from "react-hook-form";
+import {ConfigurationContext} from "../../../context/ConfigurationContext";
+import {HStack, Tooltip} from "@navikt/ds-react";
+import {ExclamationmarkTriangleFillIcon} from '@navikt/aksel-icons';
 
 export type Props = {
-	classes: ClassNameMap;
-	// TODO eivindmorch 24/03/2023 : Change to metadata as prop
-	// metadata?: IInstanceMetadataContent,
-	referencesForCollectionsToShow: string[];
+    classes: ClassNameMap;
+    // TODO eivindmorch 24/03/2023 : Change to metadata as prop
+    // metadata?: IInstanceMetadataContent,
+    referencesForCollectionsToShow: string[];
 };
 
 const IncomingDataComponent: React.FunctionComponent<Props> = (
-	props: Props
+    props: Props
 ) => {
-	const { t } = useTranslation("translations", {
-		keyPrefix: "pages.integrationForm.metadataPanel",
-	});
-	const { instanceElementMetadata, getAllMetadata } = useContext(
-		SourceApplicationContext
-	);
-	const [valueConvertings, setValueConvertings] = useState<[] | undefined>(
-		undefined
-	);
+    const {t} = useTranslation("translations", {keyPrefix: "pages.configuration.metadataPanel"});
+    const {
+        getInstanceElementMetadata,
+        instanceElementMetadata,
+        getAllMetadata,
+        allMetadata
+    } = useContext(SourceApplicationContext)
+    const {
+        completed
+    } = useContext(ConfigurationContext)
+    const {selectedMetadata, setSelectedMetadata,} = useContext(IntegrationContext)
+    const [valueConvertings, setValueConvertings] = useState<[] | undefined>(undefined);
+    const [version, setVersion] = React.useState<string>(selectedMetadata ? String(selectedMetadata.version) : '')
 
-	useEffect(() => {
-		ValueConvertingRepository.getValueConvertings(
-			0,
-			100,
-			"fromApplicationId",
-			"ASC",
-			false
-		)
-			.then((response) => {
-				setValueConvertings(response.data.content);
-			})
-			.catch((e) => {
-				console.log(e);
-				setValueConvertings([]);
-			});
-	}, []);
+    const availableVersions: IIntegrationMetadata[] = allMetadata ? allMetadata.filter(md => {
+        return md.sourceApplicationId === selectedMetadata?.sourceApplicationId &&
+            md.sourceApplicationIntegrationId === selectedMetadata.sourceApplicationIntegrationId
+    }) : []
 
-	useEffect(() => {
-		getAllMetadata(false);
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, []);
 
-	function findInstanceObjectCollectionMetadata(
-		metadataContent: IInstanceMetadataContent,
-		key: string
-	): IInstanceObjectCollectionMetadata | undefined {
-		const searchResultInCurrent: IInstanceObjectCollectionMetadata | undefined =
-			metadataContent.instanceObjectCollectionMetadata.find(
-				(instanceObjectCollectionMetadata: IInstanceObjectCollectionMetadata) =>
-					instanceObjectCollectionMetadata.key === key
-			);
-		if (searchResultInCurrent) {
-			return searchResultInCurrent;
-		}
-		for (const category of metadataContent.categories) {
-			const categorySearchResult:
-				| IInstanceObjectCollectionMetadata
-				| undefined = findInstanceObjectCollectionMetadata(
-				category.content,
-				key
-			);
-			if (categorySearchResult) {
-				return categorySearchResult;
-			}
-		}
-		return undefined;
-	}
+    const methods = useFormContext();
 
-	function getReferenceAndCollectionMetadata(
-		references: string[]
-	): [string, IInstanceObjectCollectionMetadata][] {
-		const referenceAndCollectionMetadata: [
-			string,
-			IInstanceObjectCollectionMetadata
-		][] = [];
-		references.forEach((reference: string) => {
-			if (isFieldReference(reference)) {
-				const key: string = extractFieldReferenceKey(reference);
-				const collectionMetadata:
-					| IInstanceObjectCollectionMetadata
-					| undefined = instanceElementMetadata
-					? findInstanceObjectCollectionMetadata(instanceElementMetadata, key)
-					: undefined;
-				if (collectionMetadata) {
-					referenceAndCollectionMetadata.push([reference, collectionMetadata]);
-				}
-			} else if (isCollectionFieldReference(reference)) {
-				const [index, key]: [number, string] =
-					extractCollectionFieldReferenceIndexAndKey(reference);
-				const collectionMetadata:
-					| IInstanceObjectCollectionMetadata
-					| undefined = findInstanceObjectCollectionMetadata(
-					referenceAndCollectionMetadata[index][1].objectMetadata,
-					key
-				);
-				if (collectionMetadata) {
-					referenceAndCollectionMetadata.push([reference, collectionMetadata]);
-				}
-			}
-		});
-		return referenceAndCollectionMetadata;
-	}
+    useEffect(() => {
+        ValueConvertingRepository.getValueConvertings(
+            0,
+            100,
+            "fromApplicationId",
+            "ASC",
+            false
+        )
+            .then((response) => {
+                setValueConvertings(response.data.content);
+            })
+            .catch((e) => {
+                console.log(e);
+                setValueConvertings([]);
+            });
+    }, []);
 
-	return (
-		<>
-			<Box
-				id={"incoming-form-panel"}
-				className={props.classes.panelContainer}
-				sx={metadataPanelSX}
-			>
-				<Box className={props.classes.row}>
-					<Typography variant={"h6"}>{t("header")}</Typography>
-					<HelpPopover popoverContent="Metadata er data fra innsendt skjema du kan bruke i konfigurasjon av utgående data" />
-				</Box>
-				{instanceElementMetadata && (
-					<Box id={"metadata-content-panel"} className={props.classes.panel}>
-						<MetadataContentComponent
-							classes={props.classes}
-							content={instanceElementMetadata}
-							keyToReferenceFunction={(key: string) =>
-								toInstanceFieldReference(key)
-							}
-						/>
-					</Box>
-				)}
-				{props.referencesForCollectionsToShow.length > 0 &&
-					getReferenceAndCollectionMetadata(
-						props.referencesForCollectionsToShow
-					).map(
-						(
-							[reference, objectCollectionMetadata]: [
-								string,
-								IInstanceObjectCollectionMetadata
-							],
-							index: number
-						) => (
-							<Box
-								key={"tagTreeCollectionValues-" + index}
-								className={props.classes.panel}
-							>
-								<ObjectCollectionMetadataContentComponent
-									classes={props.classes}
-									collectionIndex={index}
-									reference={reference}
-									objectCollectionMetadata={objectCollectionMetadata}
-								/>
-							</Box>
-						)
-					)}
-				{valueConvertings && (
-					<Box id={"value-converting-panel"} className={props.classes.panel}>
-						<Typography variant={"h6"}>Verdikonvertering</Typography>
-						{valueConvertings.map(
-							(valueConverting: IValueConverting, index: number) => {
-								return (
-									<div
-										id={"vc-tag-" + index}
-										key={"valueConvertingValue-" + index}
-										className={props.classes.tagWrapper}
-									>
-										<Tag
-											classes={props.classes}
-											value={"$vc{" + valueConverting.id.toString() + "}"}
-											tagKey={valueConverting.displayName}
-											name={valueConverting.displayName}
-											description={"$vc{" + valueConverting.id.toString() + "}"}
-											type={ValueType.VALUE_CONVERTING}
-										/>
-									</div>
-								);
-							}
-						)}
-					</Box>
-				)}
-			</Box>
-		</>
-	);
+    useEffect(() => {
+        getAllMetadata(false);
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, []);
+
+    const handleChange = (event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+        setVersion(event.target.value);
+        const version = Number(event.target.value)
+        const integrationMetadata: IIntegrationMetadata[] = availableVersions
+            .filter(metadata => metadata.version === version)
+        setSelectedMetadata(integrationMetadata[0])
+        if (integrationMetadata[0].id) {
+            methods.setValue('integrationMetadataId', Number(integrationMetadata[0].id))
+            getInstanceElementMetadata(integrationMetadata[0].id)
+        }
+    };
+
+    function findInstanceObjectCollectionMetadata(
+        metadataContent: IInstanceMetadataContent,
+        key: string
+    ): IInstanceObjectCollectionMetadata | undefined {
+        const searchResultInCurrent: IInstanceObjectCollectionMetadata | undefined =
+            metadataContent.instanceObjectCollectionMetadata.find(
+                (instanceObjectCollectionMetadata: IInstanceObjectCollectionMetadata) =>
+                    instanceObjectCollectionMetadata.key === key
+            );
+        if (searchResultInCurrent) {
+            return searchResultInCurrent;
+        }
+        for (const category of metadataContent.categories) {
+            const categorySearchResult:
+                | IInstanceObjectCollectionMetadata
+                | undefined = findInstanceObjectCollectionMetadata(
+                category.content,
+                key
+            );
+            if (categorySearchResult) {
+                return categorySearchResult;
+            }
+        }
+        return undefined;
+    }
+
+    function getReferenceAndCollectionMetadata(
+        references: string[]
+    ): [string, IInstanceObjectCollectionMetadata][] {
+        const referenceAndCollectionMetadata: [
+            string,
+            IInstanceObjectCollectionMetadata
+        ][] = [];
+        references.forEach((reference: string) => {
+            if (isFieldReference(reference)) {
+                const key: string = extractFieldReferenceKey(reference);
+                const collectionMetadata:
+                    | IInstanceObjectCollectionMetadata
+                    | undefined = instanceElementMetadata
+                    ? findInstanceObjectCollectionMetadata(instanceElementMetadata, key)
+                    : undefined;
+                if (collectionMetadata) {
+                    referenceAndCollectionMetadata.push([reference, collectionMetadata]);
+                }
+            } else if (isCollectionFieldReference(reference)) {
+                const [index, key]: [number, string] =
+                    extractCollectionFieldReferenceIndexAndKey(reference);
+                const collectionMetadata:
+                    | IInstanceObjectCollectionMetadata
+                    | undefined = findInstanceObjectCollectionMetadata(
+                    referenceAndCollectionMetadata[index][1].objectMetadata,
+                    key
+                );
+                if (collectionMetadata) {
+                    referenceAndCollectionMetadata.push([reference, collectionMetadata]);
+                }
+            }
+        });
+        return referenceAndCollectionMetadata;
+    }
+
+    return (
+        <>
+            <Box
+                id={"incoming-form-panel"}
+                className={props.classes.panelContainer}
+                sx={metadataPanelSX}
+            >
+                <HStack align={"center"} justify={"space-between"}>
+                    <HStack align={"center"}>
+                        <Typography variant={"h6"}>{t("header")}</Typography>
+                        <HelpPopover
+                            popoverContent="Metadata er data fra innsendt skjema du kan bruke i konfigurasjon av utgående data"/>
+                    </HStack>
+                    <HStack gap={"1"} align={"center"}>
+                        <TextField
+                            select
+                            autoComplete={"off"}
+                            disabled={completed}
+                            size={"small"}
+                            id="version-select"
+                            value={version}
+                            label={t('version')}
+                            onChange={(e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+                                handleChange(e)
+                            }}
+                        >
+                            {availableVersions.map((md, index) => {
+                                return <MenuItem
+                                    key={index}
+                                    value={md.version}>{t('version')} {md.version}
+                                </MenuItem>
+                            })}
+                        </TextField>
+                        {availableVersions.some(av => av.version > Number(version)) &&
+                            <Tooltip content={t('metadataWarning')}>
+                                <ExclamationmarkTriangleFillIcon color={"orange"} title="a11y-title" fontSize="1.5rem"/>
+                            </Tooltip>
+                        }
+                    </HStack>
+                </HStack>
+                {instanceElementMetadata && (
+                    <Box id={"metadata-content-panel"} className={props.classes.panel}>
+                        <MetadataContentComponent
+                            classes={props.classes}
+                            content={instanceElementMetadata}
+                            keyToReferenceFunction={(key: string) =>
+                                toInstanceFieldReference(key)
+                            }
+                        />
+                    </Box>
+                )}
+                {props.referencesForCollectionsToShow.length > 0 &&
+                    getReferenceAndCollectionMetadata(
+                        props.referencesForCollectionsToShow
+                    ).map(
+                        (
+                            [reference, objectCollectionMetadata]: [
+                                string,
+                                IInstanceObjectCollectionMetadata
+                            ],
+                            index: number
+                        ) => (
+                            <Box
+                                key={"tagTreeCollectionValues-" + index}
+                                className={props.classes.panel}
+                            >
+                                <ObjectCollectionMetadataContentComponent
+                                    classes={props.classes}
+                                    collectionIndex={index}
+                                    reference={reference}
+                                    objectCollectionMetadata={objectCollectionMetadata}
+                                />
+                            </Box>
+                        )
+                    )}
+                {valueConvertings && (
+                    <Box id={"value-converting-panel"} className={props.classes.panel}>
+                        <Typography variant={"h6"}>Verdikonvertering</Typography>
+                        {valueConvertings.map(
+                            (valueConverting: IValueConverting, index: number) => {
+                                return (
+                                    <div
+                                        id={"vc-tag-" + index}
+                                        key={"valueConvertingValue-" + index}
+                                        className={props.classes.tagWrapper}
+                                    >
+                                        <Tag
+                                            classes={props.classes}
+                                            value={"$vc{" + valueConverting.id.toString() + "}"}
+                                            tagKey={valueConverting.displayName}
+                                            name={valueConverting.displayName}
+                                            description={"$vc{" + valueConverting.id.toString() + "}"}
+                                            type={ValueType.VALUE_CONVERTING}
+                                        />
+                                    </div>
+                                );
+                            }
+                        )}
+                    </Box>
+                )}
+            </Box>
+        </>
+    );
 };
 
 export default IncomingDataComponent;
